Update telemetry prompts to be more informative, Add prompt to cleanup
The original telemetry opt-in prompt shown at initial configuration was a "little scary". Furthermore, it did not mention that data-sending is done anonymously.

A prompt is added to script cleanup which explains to the user that a data file exists. The prompt also includes a brief description of what the data file includes. If the user had chosen to send data, we now ask that they confirm they still wish to do so. In the case that the user now denies to send data, or they never wished to send data; we explain that they can still c
an contribute their data by pasting it at a url.

Note: The script does not currently send data anywhere. URL's will be filled in when the system to recieve data is in place.

Resolves: #3